+++ android/app/src/main/java/com/psinder/myapplication/ui/chat/ChatFragment.kt
@@ -1,10 +1,11 @@
-package com.psinder.myapplication
+package com.psinder.myapplication.ui.chat
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.psinder.myapplication.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
