+++ android/app/src/main/java/com/psinder/myapplication/ui/profile/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.psinder.myapplication
+package com.psinder.myapplication.ui.profile
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
-import androidx.navigation.findNavController
+import com.psinder.myapplication.R
 import com.psinder.myapplication.databinding.FragmentProfileBinding
 
 class ProfileFragment : Fragment() {
