+++ android/app/src/main/java/com/psinder/myapplication/ui/registration/RegistrationFragment.kt
@@ -1,20 +1,16 @@
-package com.psinder.myapplication
+package com.psinder.myapplication.ui.registration
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.view.WindowManager
-import android.widget.EditText
-import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.findNavController
-import br.com.simplepass.loading_button_lib.customViews.CircularProgressButton
+import com.psinder.myapplication.R
 import com.psinder.myapplication.databinding.FragmentRegistrationBinding
 import com.psinder.myapplication.network.RegisterData
 import com.psinder.myapplication.network.ResultWrapper
