+++ android/app/src/main/java/com/psinder/myapplication/util/BindingUtils.kt
